requirements.txt
amqp==2.2.1
beautifulsoup4==4.6.0
billiard==3.5.0.3
celery==4.0.2
certifi==2017.4.17
chardet==3.0.4
Django==1.11.3
django-cors-headers==2.1.0
djangorestframework==3.6.3
idna==2.5
kombu==4.0.2
MechanicalSoup==0.7.0
psycopg2==2.7.1
pytz==2017.2
redis==2.10.5
requests==2.18.1
six==1.10.0
urllib3==1.21.1
vine==1.1.4
